Fix issue when getting multiple things in one run
If a route required both channel and playlist data for example, the
filters from the channel request would be left behind for the next
one - in this case playlists. That results in the API getting confused,
which is not good. I fix it by resetting filters before each new request.